Implement Postive GET User Test
Implements status code, payload, idempotency and minimum performance
positive tests for GET method to /public-api/users resource.

An empty errors list is created at the start of each test. Whenever an
error occurs, the error's description is appended to the list. If the
error array is not empty by the end of the test, the test is failed and
the error array is displayed.

The verification file is created and contains helpful functions that
may be used by different tests and abstract problems.

Caching folders are removed.