Reorganizes Testing Scope
The negative tests are combined into one file. Dividing into valid and
invalid input did not do much for the clarity of the tests and made
tests harder to find.

Testing for /public-api/users/123 was replaced for general testing of
/public-api/users subordinates as we cannot ensure /public-api/users/123
always exists.

Added function to ensure a resource is empty prior to testing.