Implements Negative Tests for POST
This tests different failure states for the POST method. Namely
unauthorized access due to an invalid token, empty payload, wrong
datatype in payload and duplicate POST requests.

It also renames `basic` tests to `integrated` to be more descriptive.
It describes malformed JSON errors from as Server errors instead of
Payload errors and it is caused by the server not the sent payload.

Finally, it adds some clarifying commets to POST tests.